+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 Unreleased changes, in reverse chronological order. New entries are added at the top of this list.
 
+- [#1567](https://github.com/Zilliqa/zq2/pull/1567): Add `GetNumDSBlocks` API.
+- [#1557](https://github.com/Zilliqa/zq2/pull/1557): Add `GetTxnBodiesForTxBlockEx` API.
+- [#1534](https://github.com/Zilliqa/zq2/pull/1534): Accept `input` or `data` for transaction request payloads.
 - [#1530](https://github.com/Zilliqa/zq2/pull/1530): State at checkpoint should contain current block data as well as parent block data
 - [#1449](https://github.com/Zilliqa/zq2/pull/1449): Restructure how blocks are stored to prevent database inconsistencies.
 - [#1472](https://github.com/Zilliqa/zq2/pull/1472): Add `GetTransactionsForTxBlockEx` API.